fix Booking Status Repository

+++ src/main/java/com/aston/rapidride/repository/BookingStatusRepository.java
@@ -3,5 +3,5 @@ package com.aston.rapidride.repository;
 import com.aston.rapidride.entity.BookingStatus;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface BookingStatusRepository extends JpaRepository<BookingStatus, Integer> {
+public interface BookingStatusRepository extends JpaRepository<BookingStatus, Long> {
 }
